pyproject.toml
[project]
name = "ozempic-seguro"
version = "1.3.4"
description = "Sistema de gerenciamento de gavetas para controle seguro de medicamentos termolábeis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Caique Azevedo", email = "caiqueazevedo2005@gmail.com"}
]
requires-python = ">=3.13"
keywords = ["medicamentos", "controle", "segurança", "desktop", "customtkinter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment",
]

dependencies = [
    "customtkinter==5.2.2",
    "Pillow>=10.3,<12",
    "bcrypt==4.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "black==24.10.0",
    "ruff==0.8.4",
    "mypy==1.13.0",
]

[project.urls]
Homepage = "https://github.com/CaiqueAzevedo65/Projeto-Ozempic-Seguro"
Repository = "https://github.com/CaiqueAzevedo65/Projeto-Ozempic-Seguro.git"

[project.scripts]
ozempic-seguro = "ozempic_seguro.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/ozempic_seguro/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
fallback_version = "1.3.4"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.mypy_cache
    | build
    | dist
)/
'''

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".mypy_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # function call in default argument
    "B026",   # star-arg after keyword arg (unavoidable in CustomTkinter super().__init__)
]

[tool.ruff.lint.isort]
known-first-party = ["ozempic_seguro"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    "build/",
    "dist/",
]

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["ozempic_seguro"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src/ozempic_seguro",
    "--cov-report=html:htmlcov",
    "--cov-report=term-missing",
    "--cov-fail-under=50",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "ui: UI related tests",
]

[tool.coverage.run]
source = ["src/ozempic_seguro"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/encrypted_database.py",
    "*main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
